+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add base infrastucture for demo app in react
 - Add pricing link in README
 - Add an overview of API methods
+- Add IoT integration to device demo
 
 ### Changed
 - Prevent prebuild from increase patch number when publishing to NPM